requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.1.4
django-admin-thumbnails==0.2.6
django-crispy-forms==1.14.0
django-filter==22.1
executing==1.2.0
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
wcwidth==0.2.5
